split nested classes into separate files
