+++ time-arena-game/Assets/Assets/Scenes/GameScene.unity
@@ -2088,20 +2088,12 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: f1b8f223c584f49dca13d854546d6dd8, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  gameLength: 300
-  timeElapsedInGame: 0
-  player: {fileID: 6458975045518551228, guid: b9de9f260d9b946b8bb80f21e84fb5cd, type: 3}
-  players: []
-  gameStarted: 0
-  gameEnded: 0
-  winningTeam: 1
-  hiderSpawnPoints:
-  - {x: 4, y: -0.5, z: 0.3}
-  - {x: 7, y: -0.5, z: 3}
-  - {x: 7, y: -0.5, z: -3}
-  - {x: 9, y: -0.5, z: 0.3}
-  - {x: 13, y: 1.2, z: 0.3}
-  seekerSpawnPoint: {x: 28, y: -0.5, z: 0.3}
+  GameLength: 300
+  TimeElapsedInGame: 0
+  _players: []
+  GameStarted: 0
+  GameEnded: 0
+  WinningTeam: 1
 --- !u!4 &1944367506
 Transform:
   m_ObjectHideFlags: 0
@@ -2129,6 +2121,5 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   maxTicks: 9600
-  tps: 32
-  replayer: {fileID: 536907935286074014, guid: 343f79f26ef7d46038deaec8de340068, type: 3}
-  active: 1
+  Replayer: {fileID: 536907935286074014, guid: 343f79f26ef7d46038deaec8de340068, type: 3}
+  Active: 1
